Split components apart
